blockchain: Add FetchStats to UtxoBackend.
This adds the FetchStats method to the UtxoBackend interface and
LevelDbUtxoBackend implementation.

A summary of the changes is as follows:
- Add FetchStats method to UtxoBackend interface
- Add FetchStats method to LevelDbUtxoBackend
- Move dbFetchUtxoStats to LevelDbUtxoBackend
- Add FetchStats to the UtxoCache so that the stats can be fetched
  through the cache
