+++ blockchain/utxoio.go
@@ -11,7 +11,6 @@ import (
 	"sync"
 	"time"
 
-	"github.com/decred/dcrd/blockchain/standalone/v2"
 	"github.com/decred/dcrd/chaincfg/chainhash"
 	"github.com/decred/dcrd/database/v2"
 	"github.com/decred/dcrd/wire"
@@ -385,62 +384,6 @@ func deserializeUtxoEntry(serialized []byte, txOutIndex uint32) (*UtxoEntry, err
 	}
 
 	return entry, nil
-}
-
-// dbFetchUxtoStats fetches statistics on the current unspent transaction output
-// set.
-func dbFetchUtxoStats(dbTx database.Tx) (*UtxoStats, error) {
-	utxoBucket := dbTx.Metadata().Bucket(utxoSetBucketName)
-
-	var stats UtxoStats
-	transactions := make(map[chainhash.Hash]struct{})
-	leaves := make([]chainhash.Hash, 0)
-	cursor := utxoBucket.Cursor()
-
-	for ok := cursor.First(); ok; ok = cursor.Next() {
-		key := cursor.Key()
-		var outpoint wire.OutPoint
-		err := decodeOutpointKey(key, &outpoint)
-		if err != nil {
-			return nil, err
-		}
-
-		serializedUtxo := cursor.Value()
-		entrySize := len(serializedUtxo)
-
-		// A non-nil zero-length entry means there is an entry in the database for a
-		// spent transaction output which should never be the case.
-		if entrySize == 0 {
-			return nil, AssertError(fmt.Sprintf("database contains entry for spent "+
-				"tx output %v", outpoint))
-		}
-
-		stats.Utxos++
-		stats.Size += int64(entrySize)
-		transactions[outpoint.Hash] = struct{}{}
-
-		leaves = append(leaves, chainhash.HashH(serializedUtxo))
-
-		// Deserialize the utxo entry.
-		entry, err := deserializeUtxoEntry(serializedUtxo, outpoint.Index)
-		if err != nil {
-			// Ensure any deserialization errors are returned as database corruption
-			// errors.
-			if isDeserializeErr(err) {
-				str := fmt.Sprintf("corrupt utxo entry for %v: %v", outpoint, err)
-				return nil, makeDbErr(database.ErrCorruption, str)
-			}
-
-			return nil, err
-		}
-
-		stats.Total += entry.amount
-	}
-
-	stats.SerializedHash = standalone.CalcMerkleRootInPlace(leaves)
-	stats.Transactions = int64(len(transactions))
-
-	return &stats, nil
 }
 
 // -----------------------------------------------------------------------------
